Backport https://review.openstack.org/#/c/61037/
Start instances using fixed network when possible

Uses the fixed_network_name option to select which network
to boot test instances on.  This allows Tempest to be run
in environments where there are multiple networks available.

This does not work with nova-network because from what I can see
there is no way to retrieve the network list right now, but it
will gracefully fall back to the old behavior with nova-network
so I believe that can be left as a future enhancement if someone
needs that functionality.

Change-Id: Ia78582cac3790653c2281a5b63d953cd46d5c290
Closes-Bug: #1250866